Overlay fix and test driver changes
