+++ deployments/kubernetes/base/compute.yaml
@@ -53,7 +53,6 @@ metadata:
   name: merak-compute
   namespace: merak
 spec:
-  serviceAccountName: merak-compute
   replicas: 1
   selector:
     matchLabels:
@@ -63,6 +62,7 @@ spec:
       labels:
         app: merak-compute
     spec:
+      serviceAccountName: merak-compute
       tolerations:
       - key: "node-role.kubernetes.io/master"
         operator: "Exists"
@@ -86,7 +86,7 @@ spec:
         - name: CONCURRENCY
           value: "100"
         - name: WORKER_IMAGE
-          value: meraksim/compute-vm-worker:test
+          value: meraksim/merak-compute-vm-worker:test
         - name: NODE_NAME
           valueFrom:
             fieldRef:
