+++ deployments/kubernetes/compute.test/kustomization.yaml
@@ -18,10 +18,7 @@ images:
   newTag: test
 resources:
 - test.driver.yaml
-- service.account.yaml
-replicas:
-- name: merak-compute-vm-worker
-  count: 10
+- temporal.test.yaml
 # Exclude Merak Network
 patches:
 - patch: |-
@@ -139,21 +136,24 @@ patches:
       name: sdn-controller
       namespace: merak
 
+
+
+
 # Temporal Helm Chart
-helmCharts:
-- releaseName: temporal
-  repo: https://charts.wener.tech
-  name: temporal
-  version: 0.15.1
-  valuesInline:
-    server:
-      replicatCount: 1
-    cassandra:
-      config:
-        cluster_size: 1
-    prometheus:
-      enabled: false
-    grafana:
-      enabled: false
-    elasticsearch:
-      enabled: false
+# helmCharts:
+# - releaseName: temporal
+#   repo: https://charts.wener.tech
+#   name: temporal
+#   version: 0.15.1
+#   valuesInline:
+#     server:
+#       replicatCount: 1
+#     cassandra:
+#       config:
+#         cluster_size: 1
+#     prometheus:
+#       enabled: false
+#     grafana:
+#       enabled: false
+#     elasticsearch:
+#       enabled: false
