+++ services/merak-compute/compute.go
@@ -15,6 +15,7 @@ package main
 
 import (
 	"context"
+	"errors"
 	"flag"
 	"fmt"
 	"log"
@@ -32,6 +33,11 @@ import (
 	"go.temporal.io/api/workflowservice/v1"
 	"go.temporal.io/sdk/client"
 	"google.golang.org/grpc"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
 )
 
 var (
@@ -103,7 +109,7 @@ func main() {
 	}
 	log.Println("Successfully connected to Redis!")
 	defer handler.RedisClient.Close()
-
+	createWorkers()
 	//Start gRPC Server
 	flag.Parse()
 	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", *Port))
@@ -119,4 +125,132 @@ func main() {
 		log.Fatalf("failed to serve: %v", err)
 	}
 
+}
+
+func createWorkers() {
+	config, err := rest.InClusterConfig()
+	if err != nil {
+		log.Fatalf("Failed to get in cluster config!: %v\n", err)
+	}
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		log.Fatalf("Failed to create kube client!: %v\n", err.Error())
+	}
+	// Get all nodes
+	nodes, err := clientset.CoreV1().Nodes().List(context.Background(), metav1.ListOptions{})
+	if err != nil {
+		log.Fatalln("Failed to get nodes in the cluster!")
+	}
+	log.Println("Creating workers for nodes: ", nodes.Items)
+
+	// Label the node that this pod is running on
+	node, err := clientset.CoreV1().Nodes().Get(context.Background(), os.Getenv("NODE_NAME"), metav1.GetOptions{})
+	if err != nil {
+		log.Fatalln("Failed to get node!")
+	}
+	log.Println("Found this host: ", node.Name)
+	labelPatch := fmt.Sprintf(`[{"op":"add","path":"/metadata/labels/%s","value":"%s" }]`, constants.KUBE_NODE_LABEL_KEY, constants.KUBE_NODE_LABEL_VAL)
+	_, err = clientset.CoreV1().Nodes().Patch(context.Background(), node.Name, types.JSONPatchType, []byte(labelPatch), metav1.PatchOptions{})
+	if err != nil {
+		log.Fatalln("Failed to label node!")
+	}
+
+	log.Println("Labeled this host: ", node.Name)
+
+	// Get ENV variables
+	temporal_addr, err := getEnv(constants.TEMPORAL_ENV)
+	if err != nil {
+		temporal_addr = constants.TEMPORAL_ADDRESS
+	}
+	log.Println("WORKER: Using Temporal address from ENV " + constants.TEMPORAL_ADDRESS)
+
+	rps, err := getEnv(constants.TEMPORAL_RPS_ENV)
+	if err != nil {
+		rps = constants.WORKER_DEFAULT_RPS
+	}
+	log.Println("WORKER: Using RPS from ENV " + rps)
+
+	concurrency, err := getEnv(constants.TEMPORAL_CONCURRENCY_ENV)
+	if err != nil {
+		concurrency = constants.WORKER_DEFAULT_CONCURRENCY
+	}
+	log.Println("WORKER: Using CONCURRENCY from ENV " + concurrency)
+
+	logLevel, err := getEnv(constants.LOG_LEVEL_ENV)
+	if err != nil {
+
+		logLevel = constants.LOG_LEVEL_DEFAULT
+	}
+	log.Println("WORKER: Using log level from ENV " + logLevel)
+
+	mode, err := getEnv(constants.MODE_ENV)
+	if err != nil {
+		mode = constants.MODE_ALCOR
+	}
+	log.Println("WORKER: Using log level from ENV " + logLevel)
+
+	image, err := getEnv(constants.WORKER_IMAGE_ENV)
+	if err != nil {
+		log.Println("Image ENV not set! Using default mode " + constants.WORKER_DEFAULT_IMAGE)
+		image = constants.WORKER_DEFAULT_IMAGE
+	}
+	log.Println("WORKER: Using image from ENV " + image)
+
+	// Create worker pods, one per node (only nodes hosting vhosts)
+	for _, host := range nodes.Items {
+		worker := &corev1.Pod{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: constants.WORKER_POD_PREFIX + host.Name,
+			},
+			Spec: corev1.PodSpec{
+				NodeSelector: map[string]string{
+					constants.KUBE_NODE_LABEL_KEY: constants.KUBE_NODE_LABEL_VAL,
+				},
+				Containers: []corev1.Container{
+					{
+						Name:            constants.WORKER_POD_PREFIX + host.Name,
+						Image:           image,
+						ImagePullPolicy: constants.POD_PULL_POLICY_ALWAYS,
+						Ports: []corev1.ContainerPort{
+							{ContainerPort: constants.PROMETHEUS_PORT},
+						},
+						Env: []corev1.EnvVar{
+							{
+								Name: constants.MODE_ENV, Value: mode,
+							},
+							{
+								Name: constants.TEMPORAL_ENV, Value: temporal_addr,
+							},
+							{
+								Name: constants.TEMPORAL_RPS_ENV, Value: rps,
+							},
+							{
+								Name: constants.TEMPORAL_CONCURRENCY_ENV, Value: concurrency,
+							},
+							{
+								Name: constants.LOG_LEVEL_ENV, Value: logLevel,
+							},
+							{
+								Name: constants.TEMPORAL_TQ_ENV, Value: host.Name,
+							},
+						},
+					},
+				},
+			},
+		}
+		_, err := clientset.CoreV1().Pods(constants.TEMPORAL_NAMESPACE).Create(context.Background(), worker, metav1.CreateOptions{})
+		if err != nil {
+			log.Fatalln("Failed to create worker pod: "+constants.WORKER_POD_PREFIX+host.Name, err.Error())
+		}
+	}
+
+}
+
+func getEnv(key string) (string, error) {
+	val, ok := os.LookupEnv(key)
+	if !ok {
+		log.Println("Temporal environment variable not set, using default address.")
+		return "", errors.New("ENV variable not set " + key)
+	}
+	return val, nil
 }
